requirements.txt
click>=8.1.0
rich>=13.0.0
colorama>=0.4.6
python-dotenv>=1.0.0
gitpython>=3.1.0
openai>=1.0.0
groq>=0.9.0
google-generativeai>=0.7.0
anthropic>=0.30.0
httpx>=0.27.0
